[RAMEN9610-19706][9610] wlbt: WLBT Kernel Driver 10.3.1.0
WLBT kernel driver release scsc_release_10.3.1.0_mx450_p_ww

SCSC-Bug-Id: Rels-3122

+++ drivers/misc/samsung/scsc/mxfwconfig.c
@@ -190,7 +190,7 @@ static void mxfwconfig_get_dram_ref(struct scsc_mx *mx, struct mxmibref *cfg_ref
 		cfg_ref->size = 0;
 	} else {
 		mif->get_mifram_ref(mif, mxfwconfig->shdram, &cfg_ref->offset);
-		cfg_ref->size = mxfwconfig->shtotal;
+		cfg_ref->size = (uint32_t)mxfwconfig->shtotal;
 	}
 
 	SCSC_TAG_INFO(MX_CFG, "cfg_ref: 0x%x, size %u\n", cfg_ref->offset, cfg_ref->size);
